requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.3
jsonpickle==3.0.4
MarkupSafe==2.1.5
peewee==3.17.3
PyJWT==2.8.0
PyMySQL==1.1.0
python-dotenv==1.0.1
Werkzeug==3.0.2
